Sever changes, see description
Fixes:
    - NoteArena is now provided memory
    - Left mouse click history is now saved correctly
    - ScnMemory is now passed by pointer
Changes:
    - DrawRect now takes two v2s
Important todos:
    - Notes are drawn wrong

+++ src/scn_intrinsics.h
@@ -20,7 +20,7 @@ SquareRoot(float Float)
 inline float
 AbsoluteValue(float Float)
 {
-    float Result = fabs(Float);
+    float Result = (float)fabs(Float);
     return Result;
 }
